fix: separate statcard into component to fix build
