+++ package.json
@@ -17,6 +17,7 @@
     "next": "14.2.15",
     "react": "^18",
     "react-dom": "^18",
+    "starknet": "^6.11.0",
     "starknetkit": "^2.3.3"
   },
   "devDependencies": {
